Displays an overtime goal during playoffs if it was scored on exactly 66th minute
The code had a bug that if an overtime goal was scored in a playoff game at minute 65,
the goal wouldn't show up in the output. This was because I initially treated 65 as a shootout
goal minute and for those, only want to show the last one.

However, since the goal was not a shootout goal, it would get discarded originally and then never printed.

Fixes #22